requirements.txt
# Streamlit (legacy demo)
streamlit
streamlit-drawable-canvas

# FastAPI Web App
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12
aiorwlock>=1.4.0

# AI/ML
numpy
pandas
opencv-python-headless
transformers
torch
Pillow
peft
safetensors
faiss-cpu

# Data
pyarrow
